requirements.txt
langchain>=0.1.0
langchain-community
streamlit>=1.25.0
gradio>=3.40.0
python-dotenv
requests
beautifulsoup4
duckduckgo-search
PyYAML
langchain-google-genai